Add global view modifier with View.routing
