+++ Sources/Echeveria/SwiftUI/NavigationLink.swift
